2nd bug fix

+++ 0x12-singly_linked_lists/lists.h
@@ -21,13 +21,9 @@ typedef struct list_s
 	struct list_s *next;
 } list_t;
 
-/**
- * Function prototypes
- */
-
 size_t print_list(const list_t *h);
 size_t list_len(const list_t *h);
-int _strlen(const char *s);
+size_t _strlen(const char *s);
 list_t *add_node(list_t **head, const char *str);
 list_t *add_node_end(list_t **head, const char *str);
 void free_list(list_t *head);
